feat: Update GraphQL types and resolvers to use Long IDs for post and comment handling

+++ src/main/java/org/amalitech/bloggingplatformspring/graphql/types/GraphQLPost.java
@@ -6,13 +6,12 @@ import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.UUID;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class GraphQLPost {
-    private Integer id;
+    private Long id;
     private String title;
     private String body;
     private String authorId;
